pyproject.toml
[tool.poetry]
name = "nlp-paper-recommender"
version = "0.1.0"
description = ""
authors = ["jacasta2 <jacasta2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.1"
requests = "^2.31.0"
elsapy = "^0.5.1"
scikit-learn = "1.3.2"
joblib = "^1.3.2"
nltk = "^3.8.1"
mittens = "^0.2"
pinecone-client = "^3.2.2"
streamlit = "^1.33.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
ipywidgets = "^8.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
